+++ variables.tf
@@ -28,6 +28,34 @@ variable "new_issue" {
   default = null
 }
 
+variable "http_5xx" {
+  type = object({
+    name                    = string
+    priority                = optional(number, 2)
+    message                 = optional(string, null)
+    service_regex           = optional(string, "*")
+    critical                = optional(number, 1)
+    critical_recovery       = optional(number, 0)
+    time_window             = optional(string, "5m")
+    additional_filter_regex = optional(string, "")
+  })
+  default = null
+}
+
+variable "http_4xx" {
+  type = object({
+    name                    = string
+    priority                = optional(number, 2)
+    message                 = optional(string, null)
+    service_regex           = optional(string, "*")
+    critical                = optional(number, 1)
+    critical_recovery       = optional(number, 0)
+    time_window             = optional(string, "5m")
+    additional_filter_regex = optional(string, "")
+  })
+  default = null
+}
+
 variable "require_full_window" {
   default = false
 }